Engine: Optimize text rendering
Previously we relied on the user to create a text object, which could certainly lead to problems of possible memory leaks and so-forth. So, now all that's needed for Renderer::DrawString() is the window, the string you want to draw, an x and y position, and finally, an optional scale and foreground color.

This was achieved by making a lookup table of all the characters we support and mapping it to a respective rectangle, that way we only ever create a fixed amount of rectangles. This means that this runs in O(1) space complexity.

See... I can write good commit messages :)